Finished Mod
I finished mod up and did some testing it seems to be doing well. Made a
few changes to output text for delete and add to match it up more to the
example stuff. Feel free to mess with it to see if there are any obvious
bugs that I missed though I did text 3 different fields and it seems to
be working.