Program 2 Done?
I finished program 2, should be everything needed, if you guys get a
chance go ahead and download it and make sure it looks good to you. I
did the error handling on the extra sales manager, they get the total
sales listed but no bonus, the president is fine and I error handled the
extra ID of 162 by making sure there was a commission percent so that
random technician isn't getting a total listed on accident then.